🐛 numpy==* and scipy==* in .travis.yml
Fix .travis.yml environments to fix build

This is a possible fix to Issue #54. Instead of setting specific
versions of numpy and scipy, this sets them to anything.

The assumption here is that the most important thing should be
testing against multiple versions of scikit-learn, and that
testing against specific versions of numpy/scipy is less
relevant to correctness.